fixed warnings for VC++

+++ src/CNStructs.hpp
@@ -28,7 +28,7 @@
 // TODO: rewrite U16toU8 & U8toU16 to not use codecvt
 
 std::string U16toU8(char16_t* src);
-int U8toU16(std::string src, char16_t* des); // returns number of char16_t that was written at des
+size_t U8toU16(std::string src, char16_t* des); // returns number of char16_t that was written at des
 uint64_t getTime();
 
 // The PROTOCOL_VERSION definition is defined by the build system.
